+++ domain/src/main/java/com/todokanai/composepracticenew/model/RemoteStorageItem.kt
@@ -5,7 +5,7 @@ data class RemoteStorageItem(
     val id: Long = 0,
     val name: String,
     val address: String,
-    val port: Long,
+    val port: Int,
     val userId: String,
     val password: String
 )
